Add GCC version check and allow to trick NVCC
GCC 4.9 is the lowest GCC version to fully support OpenMP 4.0, so that is the lowest that we should support.

NVCC doesn't like GCC versions newer than 4.9, but it can be tricked to work with newer versions. As this may introduce problems, this can only be enabled by setting the environment variable TRICK_NVCC to YES when calling the build script.